Replace SupportsInt checks with try/except
The `if isinstance(x, SupportsInt)` logic turned out to be brittle,
particularly in the case of numpy arrays, which support an `__int__()`
method, but it still raises `TypeError` if the array was non-scalar.

Using `try: ... except TypeError: ...` is a bit less self-describing,
but more robust to edge cases like this.