Ignore a new warning from `scipy.stats.mode()`
In SciPy 1.9.0, a new keyword parameter `keepdims` was added to
`scipy.stats.mode()`. A `FutureWarning` is emitted if this parameter is
not explicitly specified in order to signal a change in the default
behavior.

We can explicitly specify the parameter while still supporting older
SciPy versions, and the changed default behavior doesn't affect our
application, so let's just ignore it.