Fix chunks after upsampling
Fixes an issue that resulted in a confusing exception being raised if
the interferogram shape was not an integer multiple of the downsampling
factor. In that case, the upsampling step required additional padding so
that the upsampled array shape matched the original shape. But the way
dask handled this padding was to create a new chunk rather than
extending an existing chunk.

This caused issues downstream when performing chunked operations on the
upsampled & original arrays, since the had a different number of chunks.
The issue manifested in a cryptic error message:

```
AttributeError: 'HighLevelGraph' object has no attribute
'_all_external_keys' [...]
```

that was difficult to attribute to the original inciting cause.