Daskify (#12)
* Add dask dependency; Bump isce3 dependency

* Parallelize with dask

* Disable warnings-as-errors

* Remove unused import

* Add comments

* Remove `ntiles` and `tile_overlap` params

Tiling is now handled using Dask so these parameters were redundant.

* Appease pre-commit

* Add workaround for dask/dask#10274

* Fix ImportError with older dask versions

* Bump minimum dask and rasterio versions

Bump minimum dask version to 2022.05.1. There's some cryptic fix in
https://github.com/dask/dask/pull/8865 that is necessary for some of our
uses of `dask.array.map_blocks()` to work correctly.

Bump min rasterio to 1.3. This is a casualty of requiring isce3>=0.12.
It's probably fine to continue using older rasterio versions, but I can
no longer resolve a conda environment with an older version than 1.3.
I can no longer test older versions in CI and am therefore unwilling to
officially support them.

* Move some common helper functions to util.py

* Add wrapper around `multiscale_unwrap()`

* Add wrapper around `multiscale_unwrap()` supporting array_like inputs &
outputs.
* Define protocols `DatasetReader` and `DatasetWriter`

* Fix output file locking

... in multiprocessing & distributed scheduler cases

* Add dependency on h5py>=3

* Add utility I/O classes

* Add tests for `get_tile_dims()`

* Use `typing.Iterable` instead of `collections.abc.Iterable`

... for Python<3.9

* Clean up more PEP 585 usage

* Fix h5py version in minimum dependencies CI test

* Fix chunks after upsampling

Fixes an issue that resulted in a confusing exception being raised if
the interferogram shape was not an integer multiple of the downsampling
factor. In that case, the upsampling step required additional padding so
that the upsampled array shape matched the original shape. But the way
dask handled this padding was to create a new chunk rather than
extending an existing chunk.

This caused issues downstream when performing chunked operations on the
upsampled & original arrays, since the had a different number of chunks.
The issue manifested in a cryptic error message:

```
AttributeError: 'HighLevelGraph' object has no attribute
'_all_external_keys' [...]
```

that was difficult to attribute to the original inciting cause.

* Add missing `**kwargs` to overload signature

* Improve error messages for invalid overloads