Split LICENSE into two files (#15)
Currently, GitHub doesn't automatically recognize tophu's licenses,
presumably because the text of both licenses are appened to the
`LICENSE` file, causing that file to fail pattern matching.

Other dual-licensed software repos (e.g.
https://github.com/facebook/buck2) *do* have their licenses recognized
by GitHub. The trick seems to be to put the contents of each license in
a separate file.

This has the minor downside of making the top-level directory slightly
more crowded.