Make submodules private
Prefix submodule names with an underscore to indicate that they're not
intended to be accessed by downstream projects.

Only symbols exposed in the top-level namespace should be considered
public. I don't want to treat the locations of symbols in individual
submodules as part of the API -- I'd prefer to be able to reorganize
things in the future without causing breaking changes.

In the future, we might add public submodules (such as a `tophu.typing`
submodule for custom type annotations) to store public symbols that
don't belong in the top-level namespace. But, for now, every symbol
that's intended to be public should be importable directly from `tophu`.