Pass scratch directory to underlying unwrapper
`tophu.multiscale_unwrap()` may be passed the path of a scratch
directory in which to put its intermediate outputs. Previously, it did
not communicate this to the underlying unwrap callback functor, though,
so the unwrapper would just put its scratch outputs in the default
temporary directory.

This is undesirable because:

* The user may want control over where all scratch outputs are written
  in order to avoid writing to drives with low bandwidth or low
  available disk space.
* The user may want intermediate outputs to persist in a known location
  after unwrapping for debugging purposes.

This update causes the scratch directory location to be propagated to
the underlying unwrapper in order to address this issue.